split files for dataclasses
